Refactor namespaces and imports across Playwright test projects; replace `TUnit` with `TUnit.Core` dependency.
