+++ tests/InfiniFrameTests.Playwright.BlazorWebView/ScriptSrcImportTests.cs
@@ -2,6 +2,7 @@
 // Imports
 // ---------------------------------------------------------------------------------------------------------------------
 using InfiniFrameTests.Playwright.BlazorWebView.TestUtility;
+using InfiniFrameTests.Playwright.Tests;
 using InfiniFrameTests.Shared;
 using Microsoft.Playwright;
 using System.Text.Json;
