+++ tests/InfiniFrameTests.Playwright.WebApp.React/JavascriptInteropTests.cs
@@ -1,6 +1,7 @@
 // ---------------------------------------------------------------------------------------------------------------------
 // Imports
 // ---------------------------------------------------------------------------------------------------------------------
+using InfiniFrameTests.Playwright.Tests;
 using InfiniFrameTests.Playwright.WebApp.React.TestUtility;
 
 namespace InfiniFrameTests.Playwright.WebApp.React;
