+++ tests/InfiniFrameTests.Playwright.WebApp.React/WebviewWindowTests.cs
@@ -1,6 +1,7 @@
 // ---------------------------------------------------------------------------------------------------------------------
 // Imports
 // ---------------------------------------------------------------------------------------------------------------------
+using InfiniFrameTests.Playwright.Tests;
 using InfiniFrameTests.Playwright.WebApp.React.TestUtility;
 
 namespace InfiniFrameTests.Playwright.WebApp.React;
