+++ tests/InfiniFrameTests.Playwright.WebApp.Vue/JavascriptTests.cs
@@ -1,6 +1,7 @@
 // ---------------------------------------------------------------------------------------------------------------------
 // Imports
 // ---------------------------------------------------------------------------------------------------------------------
+using InfiniFrameTests.Playwright.Tests;
 using InfiniFrameTests.Playwright.WebApp.Vue.TestUtility;
 
 namespace InfiniFrameTests.Playwright.WebApp.Vue;
