+++ tests/InfiniFrameTests.Playwright/Tests/SharedJavascriptInteropTests.cs
@@ -5,7 +5,7 @@ using InfiniFrame;
 using InfiniFrameTests.Shared;
 using Microsoft.Playwright;
 
-namespace InfiniFrameTests.Playwright;
+namespace InfiniFrameTests.Playwright.Tests;
 // ---------------------------------------------------------------------------------------------------------------------
 // Code
 // ---------------------------------------------------------------------------------------------------------------------
